upgrade configuration, fix format string warnings

+++ HexaFlip/Source/AI/JCSNegaScoutGameAlgorithm.m
@@ -48,8 +48,8 @@
     
     float score = [self negaScoutWithDepth:_depth alpha:-INFINITY beta:INFINITY principalVariation:pv];
     
-    NSLog(@"analyzed %u nodes in %.3f seconds, got principal variation [%@] with score %.3f%@",
-          _count, [[NSDate date] timeIntervalSinceDate:start], [pv componentsJoinedByString:@", "], score, _canceled ? @" (canceled)" : @"");
+    NSLog(@"analyzed %lu nodes in %.3f seconds, got principal variation [%@] with score %.3f%@",
+          (unsigned long)_count, [[NSDate date] timeIntervalSinceDate:start], [pv componentsJoinedByString:@", "], score, _canceled ? @" (canceled)" : @"");
     
     return [pv count] > 0 ? pv[0] : nil;
 }
@@ -139,7 +139,7 @@
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"(NegaScout Algorithm; %@; Depth %u)", _heuristic, _depth];
+    return [NSString stringWithFormat:@"(NegaScout Algorithm; %@; Depth %lu)", _heuristic, (unsigned long)_depth];
 }
 
 @end
